% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskScoring.r
\name{computeExposureInfoRisk}
\alias{computeExposureInfoRisk}
\title{Compute risk}
\usage{
computeExposureInfoRisk(exposureInfo, riskConfig)
}
\description{
Compute risk
}
\examples{
computeExposureInfoRisk()
}
