% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskScoring.r
\name{generateRandomExposures}
\alias{generateRandomExposures}
\title{Randomly generate exposure data}
\usage{
generateRandomExposures(count, currentDay, version)
}
\description{
Randomly generate exposure data
}
\examples{
generateRandomExposures()
}
