% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskScoring.r
\name{getRiskConfig}
\alias{getRiskConfig}
\title{Get the default risk configuration parameters}
\usage{
getRiskConfig()
}
\description{
Get the default risk configuration parameters
}
\examples{
getRiskConfig()
}
